refactor: refactor of the process using overpass + clean old packages
