+++ .eslintignore
@@ -1,0 +1,3 @@
+# /node_modules/* and /bower_components/* in the project root are ignored by default
+
+scripts/wikipedia_dic.js
