+++ package.json
@@ -1,56 +1,39 @@
 {
   "name": "lascallesdelasmujeres_data_scripts",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "Scripts to generate data for the geochicasOSM project, #LasCallesDeLasMujeres",
+  "main": "./scripts/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "initial-step": "node ./index",
+    "initial-step": "node ./scripts/index",
     "wikipedia-step": "node ./scripts/apply_wikipedia.js",
-    "final-step": "node ./scripts/parse_final.js"
+    "final-step": "node ./scripts/parse_final.js",
+    "lint": "eslint --fix  scripts/**"
   },
   "author": "Jessica Sena",
   "license": "MIT",
   "dependencies": {
+    "@turf/bbox": "^6.0.1",
+    "@turf/bbox-polygon": "^6.0.1",
     "@turf/boolean-contains": "^6.0.1",
     "@turf/flatten": "^5.1.5",
-    "@turf/union": "^6.0.3",
-    "binary-split": "^0.1.2",
+    "@turf/square-grid": "^6.0.2",
     "fs": "0.0.1-security",
     "fuse.js": "^3.2.0",
-    "geojson-flatten": "^0.2.2",
-    "geojson-normalize": "0.0.1",
-    "join-stream": "0.0.0",
     "line-by-line": "^0.1.6",
-    "lineclip": "^1.1.5",
-    "minimist": "^1.2.0",
-    "mkdirp": "^0.5.1",
-    "npm-run-all": "^4.1.5",
-    "progress": "^1.1.8",
     "query-overpass": "^1.5.5",
-    "request": "^2.67.0",
-    "shapefile": "^0.3.0",
-    "shapefile-stream": "0.0.4",
     "split": "^1.0.0",
-    "through2": "^2.0.0",
-    "tile-cover": "^3.0.1",
-    "tile-reduce": "^3.0.0",
-    "tilebelt": "^1.0.1",
-    "turf": "^3.0.14",
-    "turf-extent": "^1.0.4",
-    "turf-line-distance": "^1.0.2",
-    "turf-linestring": "^1.0.2",
-    "underscore": "^1.8.3",
-    "unzip2": "^0.2.5",
-    "which-polygon": "^2.1.0",
     "yargs": "^11.0.0"
   },
   "devDependencies": {
-    "@turf/bbox": "^6.0.1",
-    "@turf/bbox-polygon": "^6.0.1",
-    "@turf/square-grid": "^6.0.2",
+    "babel-eslint": "^10.1.0",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-prettier": "^3.1.3"
+    "eslint-plugin-prettier": "^3.1.3",
+    "husky": "^4.2.5"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run lint"
+    }
   }
 }
