+++ scripts/filters.js
@@ -1,0 +1,161 @@
+const filters = {
+  es: {
+    filter01: [
+      'Paseo',
+      'Río',
+      'Avenida',
+      'Hacienda',
+      'Puerto',
+      'Callejón',
+      'Calle',
+      'Calzada',
+      'Camino',
+      'Av.',
+      'Paso',
+      'Cañada',
+      'Minas',
+      'Cerrada',
+      'Puebla',
+      'Principal',
+      'Central',
+      'Primera',
+      'Segunda',
+      'Portón',
+      'Lateral',
+      'Calz.',
+      'Corrido',
+      'Casa',
+      'Villa',
+      'Mejía',
+      'Vía',
+      'Via',
+      'Real',
+      'Isla',
+      'Avendida',
+      'Marisma',
+      'Rada',
+      'Raudal',
+      'Ribera',
+      'Embocadura',
+      'Cataratas',
+      'Médanos',
+      'Mirador',
+      'Av',
+      'Jardín',
+      'A.',
+      'Circuito',
+      'Gral.',
+      'Rincón',
+      'Calz',
+      'Rinconada',
+      'Periférico',
+      'Cda',
+      'Jardin',
+      'C.',
+      'Callejon',
+      'Colegio',
+      'Valle',
+      'avenida',
+      'camino',
+      'calle',
+      'Calle',
+      'Rotonda',
+      'Parqueo',
+      'Parque',
+      'entrada',
+      'Entrada',
+      'sendero',
+      'Sendero',
+      'Pasaje',
+      'pasaje',
+      'Puerto',
+      'Ciudad',
+      'Puente',
+      'Boulevard',
+      'Agrosuperior',
+      'Bodegas',
+      'Autobanco',
+      'SkyTrace',
+      'Plaza',
+      'Motel',
+      'C/',
+      'Rotonda',
+      'Drive',
+      'Residencial',
+      'Automac',
+      'Auto',
+      'Transcersal',
+      'Inter',
+      'Pasillo',
+      'Centro',
+      'Caminito',
+      'Arandas',
+      'Proveedores',
+      'Cajero',
+      'Zona',
+      'Primer',
+      'Res.',
+    ],
+    filter02: ['de las ', 'de la ', 'de los ', 'de lo ', 'del ', 'de ', 'en '],
+  },
+  ca: {
+    filter01: [
+      'Avinguda',
+      'avinguda',
+      'Túnel',
+      'Camí',
+      'Carrer',
+      'Ctra.',
+      'Passatge',
+      'Ronda',
+      'Passeig',
+      'Viaducte',
+      'Via',
+      'carrer',
+      'Torrent',
+      'camí',
+      'Rotonda',
+      'Plaça',
+      'Jardins',
+      'Jardí',
+      'Parc',
+      'Accés',
+      'Baixada',
+      'Rambla',
+      'Travessera',
+      'travessera',
+      'Riera',
+      'plaça',
+      'Gran',
+      'Passage',
+      'Placeta',
+      'Sant',
+      'antiga',
+      'Ptge.',
+      'Pont',
+      'Travessia',
+      'la',
+      'Cerrer',
+      'Pla',
+      'Marquès',
+      'Av.',
+      'Antic',
+      'Cami',
+      'sendero',
+      'entrada',
+      'avinguda',
+      'cami',
+      'passeig',
+      'Nostra',
+      'passatge',
+      'Pista',
+      'Corriol',
+      'Costa',
+    ],
+    filter02: ['de la ', 'de l\'', 'de les ', 'dels ', 'del ', 'de ', 'd\''],
+  },
+};
+
+module.exports = {
+  filters,
+};
